🔨 Refactor: 타입을 별도의 파일로 분리
