+++ src/components/table/cells/UrlCell.tsx
@@ -1,9 +1,5 @@
 import { memo } from 'react';
 
-interface UrlCellProps {
-  url: string;
-}
-
 const UrlCell = memo(({ url }: UrlCellProps) => {
   return <span className='text-blue-500 user-select-none'>{url}</span>;
 });
