+++ src/pages/service-dashboard/ServiceDashboard.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import StatCardGrid from './components/StatCardGrid';
-import { IconType } from '@/types/service-dashboard/stat-card.type';
 import ServiceReport from './components/ServiceReport';
 import DashboardTitle from '@/components/DashboardTitle';
 import ServiceChart from './components/ServiceChart';
